backend/requirements.txt
fastapi
uvicorn
streamlit
llama-index
# llama-index-llms-openai
# llama-index-embeddings-openai
llama-index-llms-google-genai
llama-index-embeddings-gemini
google-generativeai
sentence-transformers
transformers
python-dotenv
pdfplumber
llama-index-readers-file
pymupdf
google-api-core
pytesseract
pdf2image
Pillow